moved the files around and renamed and combined them to make the the files more streamlined. Still can't get it to work. Left and right buttons are now in one file - it was throwing an error because app was defined in two places when both files were linked in the html.
