--- Btn-Modules/LftBtn
@@ -1,23 +1,0 @@
-const App = {
-    init() {
-        this.cacheDom();
-        this.bindEventListerners();
-        this.render();
-    },
-    cacheDom() {
-        this.root = document.querySelector('#App');
-        this.box = this.root.querySelector('.box');
-        this.leftBtn = this.root.querySelector('.leftBtn');
-    },
-    bindEventListeners() {
-        this.leftBtn.addEventListener('click', this.----.bind(this));
-        
-    },
-    leftClick() {
-
-    },
-    render() {}
-};
-
-App.init();
-window.App = App;
