--- Btn-Modules/RtBtn
@@ -1,22 +1,0 @@
-const App = {
-    init() {
-        this.cacheDom();
-        this.bindEventListerners();
-        this.render();
-    },
-    cacheDom() {
-        this.root = document.querySelector('#App');
-        this.box = this.root.querySelector('.box');
-        this.rightBtn = this.root.querySelector('.rightBtn');
-    },
-    bindEventListeners() {
-        this.rightBtn.addEventListener('click', this.----.bind(this));
-    },
-    rightClick() {
-        
-    },
-    render() {}
-};
-
-App.init();
-window.App = App;
