+++ index.html
@@ -18,6 +18,6 @@
                     <button class="rightB btn">&#x261B;</button>
                 </div>
              </div>
-        <script src="Btn-Modules/RtBtn.js"></script> 
+        <script src="main.js"></script> 
     </body>
 </html>
