+++ style.css
@@ -14,6 +14,7 @@ title {
 }
 
 .box {
+    border: 2px dashed black;
     width: 300px;
     height: 200px;
     position: fixed;
